requirements.txt
configparser>=3.7.4
loguru>=0.3.1
Pillow>=6.1.0
schedule>=0.6.0
peewee >=3.10.0